+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## [0.0.3]
+
+- Fix report referer formatting
+- Rename "external links" to "outbound links"
+- Validate command-line URL
+
+
 ## [0.0.2]
 
 - Improve error handling